Conflicts:
	cpu/avr/radio/rf230bb/halbb.c
	cpu/avr/radio/rf230bb/rf230bb.c
	examples/udp-ipv6/udp-client.c
	platform/avr-atmega128rfa1/contiki-conf.h
	tools/cooja/apps/avrora/cooja.config
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/AvrCompileDialog.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/AvrMoteMemory.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/MicaZMote.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/MicaZMoteType.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/interfaces/AvroraUsart0.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/interfaces/MicaZLED.java
	tools/cooja/apps/avrora/src/se/sics/cooja/avrmote/interfaces/MicaZRadio.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/MspMote.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/MspMoteType.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/interfaces/CC2420RadioPacketConverter.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/interfaces/Msp802154Radio.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/plugins/MspBreakpoint.java
	tools/cooja/apps/mspsim/src/se/sics/cooja/mspmote/plugins/MspCodeWatcher.java
	tools/cooja/java/se/sics/cooja/Watchpoint.java
	tools/cooja/java/se/sics/cooja/interfaces/CustomDataRadio.java
	tools/cooja/java/se/sics/cooja/plugins/BreakpointsUI.java
	tools/cooja/java/se/sics/cooja/plugins/CodeUI.java
	tools/cooja/java/se/sics/cooja/radiomediums/AbstractRadioMedium.java
	tools/mspsim

+++ core/net/mac/frame802154.h
@@ -54,7 +54,6 @@
  *  This file converts to and from a structure to a packed 802.15.4
  *  frame.
  *
- *    $Id: frame802154.h,v 1.3 2010/02/18 21:00:28 adamdunkels Exp $
 */
 
 
@@ -156,14 +155,14 @@ typedef struct {
   uint8_t src_addr[8];      /**< Source address */
   frame802154_aux_hdr_t aux_hdr;    /**< Aux security header */
   uint8_t *payload;     /**< Pointer to 802.15.4 frame payload */
-  uint8_t payload_len;  /**< Length of payload field */
+  int payload_len;  /**< Length of payload field */
 } frame802154_t;
 
 /* Prototypes */
 
-uint8_t frame802154_hdrlen(frame802154_t *p);
-uint8_t frame802154_create(frame802154_t *p, uint8_t *buf, uint8_t buf_len);
-uint8_t frame802154_parse(uint8_t *data, uint8_t length, frame802154_t *pf);
+int frame802154_hdrlen(frame802154_t *p);
+int frame802154_create(frame802154_t *p, uint8_t *buf, int buf_len);
+int frame802154_parse(uint8_t *data, int length, frame802154_t *pf);
 
 /** @} */
 #endif /* FRAME_802154_H */
